rust/hpet: Clean up type mismatch for num_timers property
Now `num_timers` is `usize` other than `u8`. Although the type field in
`declare_properties` macro hasn't been used, it's better to explicitly
point this out and clean up this before doing other property work.

Link: https://lore.kernel.org/r/20250920160520.3699591-11-zhao1.liu@intel.com

+++ rust/hw/timer/hpet/src/device.rs
@@ -909,7 +909,7 @@ declare_properties! {
         HPETState,
         num_timers,
         unsafe { &qdev_prop_usize },
-        u8,
+        usize,
         default = HPET_MIN_TIMERS
     ),
     define_property!(
